chore: rename base personal item
