+++ app/javascript/dashboard/routes/dashboard/settings/personal/BasePersonalItem.vue
